backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pandas
websockets
httpx
